--- eslint/babel-eslint-parser/src/parse.js
@@ -1,23 +1,0 @@
-import { parseSync as babelParse } from "@babel/core";
-import convert from "./convert";
-import { normalizeBabelParseConfig } from "./configuration";
-
-export default function parse(code, options) {
-  const parseOptions = normalizeBabelParseConfig(options);
-  let ast;
-
-  try {
-    ast = babelParse(code, parseOptions);
-  } catch (err) {
-    if (err instanceof SyntaxError) {
-      err.lineNumber = err.loc.line;
-      err.column = err.loc.column;
-    }
-
-    throw err;
-  }
-
-  convert(ast, code);
-
-  return ast;
-}
